backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.4.3",
    "pymupdf>=1.26.4",
    "servestatic>=3.1.0",
    # Uvicorn and its dependencies
    "uvicorn>=0.37.0",
    # "uvloop>=0.15.1; sys_platform != 'win32' and (sys_platform != 'cygwin' and platform_python_implementation != 'PyPy')",
    "colorama>=0.4; sys_platform == 'win32'",
    "httptools>=0.7.1",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
    "django-dynamic-filenames>=1.4.0",
]

[dependency-groups]
dev = ["django-stubs>=5.2.5", "ruff>=0.13.3"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]
